+++ src/main/java/com/sandymandy/pleasurecraft/util/Utils.java
@@ -30,6 +30,8 @@ import java.util.*;
 
 public class Utils {
 
+
+
     public static BlockPos getBlockPosFromVec3d(Vec3d pos) {return new BlockPos((int) pos.getX(), (int) pos.getY(), (int) pos.getZ());}
 
     public static Settlement findNearestSettlement(World world, BlockPos pos) {